Rename module `edit_command` to `rpc`
