+++ rust/src/rpc.rs
